Restructuring of src/filter.cu

+++ src/filter.h
@@ -4,7 +4,7 @@
 #ifdef __CUDACC__
 extern "C" {
 #endif
-void box_blur(const uint8_t *, uint8_t *, const int, const int);
+void box_blur(const uint8_t *, uint8_t *, const size_t, const size_t);
 #ifdef __CUDACC__
 } // extern "C"
 #endif
